\name{clusters.Flu2.csv}
\docType{data}
\alias{clusters2.Flu.csv}
\title{Cluster structure example input/output file for Beast 2}
\description{
  This is an example of a cluster structure input/output file for package function RandomCluster2.
  }
\usage{clusters.Flu.csv}
\format{CSV file}
\source{derived from BEAST v1.7 distribution package}
\references{
  Drummond AJ, Suchard MA, Xie D & Rambaut A (2012) Bayesian phylogenetics with BEAUti and the BEAST 1.7.
  Molecular Biology And Evolution 29: 1969-1973.
}
\keyword{datasets}
\keyword{BEAST Software}
\keyword{phylogenetics}