\name{Flu2.xml}
\docType{data}
\alias{Flu2.xml}
\title{Example BEAST 2 input file}
\description{
  This is an example input file for BEAST 1 with tip dates.
  This file is provided as example with BEAST v1.8.2}
\usage{Flu2.xml}
\format{XML file}
\source{derived from BEAST v1.7 distribution package}
\references{
  Drummond AJ, Suchard MA, Xie D & Rambaut A (2012) Bayesian phylogenetics with BEAUti and the BEAST 1.7.
  Molecular Biology And Evolution 29: 1969-1973.
}
\keyword{datasets}
\keyword{BEAST Software}
\keyword{phylogenetics}