\name{Flu_BEAST_232.clusters.Flu.csv}
\docType{data}
\alias{Flu_BEAST_232.clusters.Flu.csv}
\title{Cluster structure example input file for the RandomCluster function}
\description{
  This is an example of a cluster structure input file for package function RandomCluster.
  }
\usage{Flu_BEAST_232.clusters.Flu.csv}
\format{CSV file}
\source{derived from BEAST v1.7 distribution package}
\references{
  Drummond AJ, Suchard MA, Xie D & Rambaut A (2012) Bayesian phylogenetics with BEAUti and the BEAST 1.7.
  Molecular Biology And Evolution 29: 1969-1973.
}
\keyword{datasets}
\keyword{BEAST Software}
\keyword{phylogenetics}